requirements.txt
numpy==1.18.0
torch==1.6.0
gym==0.17.3
atari-py==0.2.6
box2d-py==2.3.8
scipy==1.5.4
opencv-python==4.4.0.46
addict==2.4.0
tqdm

# for distributed learning
redis==3.5.3  # for ray
ray==1.3.0
pyzmq==20.0.0
pyarrow==3.0.0

# for log
six>=1.13.0
wandb==0.10.11
matplotlib==3.3.3
plotly==4.13.0
